move HiringFeeComponent to Conversations namespace
