+++ app/views/conversations/show.html.erb
@@ -24,7 +24,7 @@
     <%= render @conversation.messages %>
   </ul>
 
-  <%= render HiringFeeComponent.new(current_user, @conversation) %>
+  <%= render Conversations::HiringFeeComponent.new(current_user, @conversation) %>
 
   <div class="mt-16 max-w-prose w-full mx-auto px-4 sm:px-0 pb-6">
     <% if policy(@message).create? %>
